VoronoiSpheres

Documentation for VoronoiSpheres.

VoronoiSpheres.AbstractDomain

VoronoiSpheres.FDDomain

VoronoiSpheres.HVLayout

VoronoiSpheres.HyperDiffusion

VoronoiSpheres.MassCoordinate

VoronoiSpheres.PressureCoordinate

VoronoiSpheres.Shell

VoronoiSpheres.SigmaCoordinate

VoronoiSpheres.SpectralDomain

VoronoiSpheres.SpectralSphere

VoronoiSpheres.Stencils.Fix

VoronoiSpheres.VHLayout

VoronoiSpheres.VerticalCoordinate

VoronoiSpheres.Stencils.TRiSK

VoronoiSpheres.Stencils.average_ev_form

VoronoiSpheres.Stencils.average_ie

VoronoiSpheres.Stencils.average_iv

VoronoiSpheres.Stencils.average_iv_form

VoronoiSpheres.Stencils.average_ve

VoronoiSpheres.Stencils.average_vi_form

VoronoiSpheres.Stencils.centered_flux

VoronoiSpheres.Stencils.contraction

VoronoiSpheres.Stencils.cross_product

VoronoiSpheres.Stencils.curl

VoronoiSpheres.Stencils.div_centered_flux

VoronoiSpheres.Stencils.div_form

VoronoiSpheres.Stencils.divergence

VoronoiSpheres.Stencils.dot_grad

VoronoiSpheres.Stencils.dot_prod_contra

VoronoiSpheres.Stencils.dot_product

VoronoiSpheres.Stencils.dot_product_form

VoronoiSpheres.Stencils.grad_form

VoronoiSpheres.Stencils.gradient

VoronoiSpheres.Stencils.gradient3d

VoronoiSpheres.Stencils.gradperp

VoronoiSpheres.Stencils.mul_grad

VoronoiSpheres.Stencils.perp

VoronoiSpheres.Stencils.squared_covector

VoronoiSpheres.VerticalInterpolation.interpolate!

VoronoiSpheres.VoronoiOperators.AsDensity

VoronoiSpheres.VoronoiOperators.pdv

VoronoiSpheres.ZeroArrays.zero_array

VoronoiSpheres.allocate_field

VoronoiSpheres.allocate_fields

VoronoiSpheres.data_layout

VoronoiSpheres.hyperdiff_shell!

VoronoiSpheres.hyperdiffusion!

VoronoiSpheres.laplace_dx

VoronoiSpheres.mass_coordinate

VoronoiSpheres.mass_coordinate

VoronoiSpheres.mass_level

VoronoiSpheres.periodize!

VoronoiSpheres.pressure_level

VoronoiSpheres.scratch_hyperdiff

VoronoiSpheres.scratch_space

VoronoiSpheres.shell

VoronoiSpheres.transpose!

General

VoronoiSpheres.AbstractDomain — Type

Parent type of SpectralDomain and FDDomain

VoronoiSpheres.FDDomain — Type

FDDomain <: AbstractDomain`

VoronoiSpheres.HVLayout — Type

struct HVLayout{rank} end
layout = HVLayout(rank)

Singleton type describing a multi-layer data layout where horizontal layers are contiguous in memory. rank is the number of horizontal indices (1 or 2).

VoronoiSpheres.HyperDiffusion — Method

filter = HyperDiffusion(fieldtype::Symbol, domain, niter::Int, nu)                  # user-friendly
filter = HyperDiffusion{fieldtype, D, F, X}(domain::D, niter::Int, nu::F, extra::X) # internal

Return a filter that applies Laplacian diffusion iterated niter times with hyperdiffusive coefficient nu ≥ 0 on fields of type fieldtype. Supported field types are :scalar, :vector, :vector_curl, :vector_div.

The filter is to be used as:

# out-of-place, allocates `scratch`, allocates and returns `filtered_data`
filtered_data = filter(void, in_data, void)

# mutating, non-allocating
out_data = filter(out_data, in_data, scratch)

# in-place, non-allocating
inout_data = filter(inout_data, inout_data, scratch)

filter(out_data, in_data, scratch) returns hyperdiffusion!(filter.domain, filter, out_data, in_data, scratch). If scratch::Void, then

scratch = scratch_space(filter, in_data)

The user-friendly HyperDiffusion may be specialized for specific types of domain. The specialized method should call the internal constructor which takes an extra argument extra, stored as filter.extra for later use by hyperdiffusion!. The default user-friendly constructor sets extra=nothing.

Similarly, scratch_space(filter, in_data) calls scratch_hyperdiff(f.domain, Val(fieldtype), in_data) which by default returns nothing. scratch_hyperdiff may be specialized for specific domain types and fieldtype.

VoronoiSpheres.MassCoordinate — Type

abstract type MassCoordinate{N} <: VerticalCoordinate{N} end

Parent type for a mass-based vertical coordinate. Children types should specialize mass_level. See also VerticalCoordinate and mass_coordinate.

VoronoiSpheres.PressureCoordinate — Type

abstract type PressureCoordinate{N} <: VerticalCoordinate{N} end

Parent type for a pressure-based vertical coordinate. Children types should specialize pressure_level and mass_level. See also VerticalCoordinate.

VoronoiSpheres.Shell — Type

multi_layer_domain = Shell(nz::Int, layer::AbstractDomain, layout)

Return a multi-layer domain made of nz layers with data layout specified by layout. Unless you know what you are doing, it is recommended to use rather:

multi_layer_domain = shell(nz::Int, layer::AbstractDomain)

which gets the data layout from data_layout(layer). Otherwise, multi_layer_domain may be non-optimal or non-usable.

VoronoiSpheres.SigmaCoordinate — Type

sigma = SigmaCoordinate(N, ptop) <: PressureCoordinate{N}

Pressure based sigma-coordinate for N levels with top pressure ptop. Pressure levels are linear in vertical coordinate k : k/N = (ps-p)/(ps-ptop) where k ranges from 0 (ground) to N (model top).

VoronoiSpheres.SpectralDomain — Type

SpectralDomain <: AbstractDomain

parent type of SpectralSphere

VoronoiSpheres.SpectralSphere — Type

Parent type for spherical domains using spherical harmonics.

VoronoiSpheres.VHLayout — Type

struct VHLayout{rank} end
layout = VHLayout(rank)

Singleton type describing a multi-layer data layout where vertical columns are contiguous. rank is the number of horizontal indices (1 or 2).

VoronoiSpheres.VerticalCoordinate — Type

abstract type VerticalCoordinate{N} end

Parent type for generalized vertical coordinates ranging from 0 to N. See also PressureCoordinate.

VoronoiSpheres.allocate_field — Function

field = allocate_field(kind::Symbol, domain::AbstractDomain, precision::Type)

Allocate a field of the given kind and precision over the given domain. Typical values for precision are Float32, Float64 or ForwardDiff.Dual. Depending on the domain, valid values for kind may include :scalar, :vector, :scalar_spec, scalar_spat (spectral/spatial representation of a scalar field), :vector_spec, vector_spat (spectral/spatial representation of a vector field).

Internally, allocate_field(kind::Symbol, domain, F) returns allocate_field(Val(kind), domain, F). To specialize allocate_field for MyDomain <: AbstractDomain, one must provide methods for:

allocate_field(::Val{kind}, domain::MyDomain, F)

where symbol kind::Symbol is one of the valid field kinds for that domain.

VoronoiSpheres.allocate_fields — Function

fields = allocate_fields(kinds::Tuple, domain::AbstractDomain, F::Type)
fields = allocate_fields(kinds::NamedTuple, domain::AbstractDomain, F::Type)

Allocate a (named) tuple of fields according to the provided kinds. For instance:

fields = allocate_fields((:vector, :scalar), domain, F)
fields = allocate_fields((a=:vector, b=:scalar), domain, F)

are equivalent to, respectively:

fields = (allocate_field(:vector, domain, F), allocate_field(:scalar, domain, F))
fields = (a=allocate_field(:vector, domain, F), b=allocate_field(:scalar, domain, F))

VoronoiSpheres.data_layout — Method

layout = data_layout(shell::Shell)

Return layout describing the data layout of multi-layer domain shell.

layout = data_layout(domain::Domain)

Return layout describing the preferred data layout for a shell made of layers of type Domain.

Typical values for layout are the singletons HVLayout() (layers are contiguous in memory) and VHLayout (columns are contiguous in memory).

VoronoiSpheres.hyperdiff_shell! — Function

new_data = hyperdiff_shell!(layer, layout, filter, out_data, in_data, scratch)

Apply hyperdiffusive filtertoindata. The result is written intooutdataand returned. Ifdomain::SpectralDomain, the filter applies to spectral coefficients. Ifout_data::Void, it is adequately allocated. SeeHyperDiffusion.

VoronoiSpheres.hyperdiffusion! — Method

new_data = hyperdiffusion!(domain, filter, out_data, in_data, scratch)

Apply hyperdiffusive filtertoindata. The result is written intooutdataand returned. Ifdomain::SpectralDomain, the filter applies to spectral coefficients. Ifout_data::Void, it is adequately allocated. SeeHyperDiffusion.

VoronoiSpheres.laplace_dx — Function

Estimates the largest eigenvalue -lambda=dx^-2 of the scalar Laplace operator and returns dx which is a (non-dimensional) length on the unit sphere characterizing the mesh resolution. By design, the Courant number for the wave equation with unit wave speed solved with time step dt is dt/dx.

VoronoiSpheres.mass_coordinate — Function

mcoord = mass_coordinate(pcoord::PressureCoordinate, metric_cov=1)

Return the mass-based coordinate deduced from pcoord and the covariant metric factor metric_cov. metric_cov can be a scalar or a vector. In the latter case metric_cov[ij]) is the metric factor at horizontal positionij. The objectmcoord` can then be used with:

m = mass_level(k, ij, masstot, mcoord)

With metric_cov==1, masstot should be in Pa ( kg/m²⋅(m/s²) ). With metric_cov in m², masstot should be in kg⋅(m/s²). m has the same unit as mass_tot. If masstot is covariant (integral over a cell) then metric_cov should include the cell area. See also mass_level.

VoronoiSpheres.mass_coordinate — Method

mcoord = mass_coordinate(mcoord::MassCoordinate)

Return mcoord itself, unchanged. Interim function for backwards compatibility.

VoronoiSpheres.mass_level — Function

m = mass_level(k, ij, masstot, mcoord::MassCoordinate{N})

Return mass m in level k/2 and at horizontal position ij, as prescribed by vertical coordinate mcoord and total mass masstot.

So-called full levels correspond to odd values k=1,2...2N-1 while interfaces between full levels (so-called half-levels) correspond to even values k=0,2...2N

masstot may be:

per unit surface, with unit Pa

per unit non-dimensional surface (e.g. on the unit sphere), with unit kg.(m/s²))

Which convention is appropriate depends on the metric_factor provided when constructing mcoord. See also mass_coordinate.

VoronoiSpheres.periodize! — Method

periodize!(data, box::AbstractBox, mgr)

Enforce horizontally-periodic boundary conditions on array data representing grid point values in box. data may also be a collection, in which case periodize! is applied to each element of the collection. Call periodize! on data obtained by computations involving horizontal averaging/differencing.

VoronoiSpheres.pressure_level — Function

p = pressure_level(k, ps, vcoord::PressureCoordinate{N})

Returns pressure p corresponding to level k/2 as prescribed by vertical coordinate vcoord and surface pressure ps.

So-called full levels correspond to odd values k=1,2...2N-1 while interfaces between full levels (so-called half-levels) correspond to even values k=0,2...2N

VoronoiSpheres.scratch_hyperdiff — Method

scratch = scratch_hyperdiff(domain, Val(fieldtype), field)

Return scratch space used to apply hyperdiffusion on domain for a field of a certain fieldtype. See HyperDiffusion and hyperdiffusion!.

VoronoiSpheres.scratch_space — Method

scratch = scratch_space(filter::AbstractFilter, field, [scratch])

If scratch is omitted or ::Void, return scratch space for applying filter to field, allocated by scratch_hyperdiff. Otherwise just return scratch. See also HyperDiffusion.

VoronoiSpheres.shell — Method

multi_layer_domain = shell(nz::Int, layer::AbstractDomain)

Return a multi-layer domain made of nz layers with data layout specified by data_layout(layer).

VoronoiSpheres.transpose! — Method

x_ji = transpose!(x_ji, mgr, x_ij)
y_ji = transpose!(void, mgr, y_ij)

Transposes x_ij and writes the result into x_ji, which may be ::Void, in which case it is allocated.

When working with shells it is sometimes useful to transpose fields for performance. VoronoiSpheres.transpose! can be specialized for specific managers, for instance:

import VoronoiSpheres: transpose!, Void
using Strided: @strided
function transpose!(x, ::MultiThread, y)
   @strided permutedims!(x, y, (2,1))
   return x # otherwise returns a StridedView
end
transpose!(::Void, ::MultiThread, y) = permutedims(y, (2,1)) # for non-ambiguity

Vertical interpolation

VoronoiSpheres.VerticalInterpolation.interpolate! — Function

interpolated = interpolate!(mgr::LoopManager, domain::Shell, field, coord, refs, increasing)

Interpolate field defined on 3D domain to reference values refs of coord. If coord increases with level number, increasing==true and vice-versa. refs must be sorted according to increasing. out may be void. Example:

T_ref = interpolate!(mgr, void, domain, [850., 500.], temperature, pressure, false)

With these arguments, interpolate! is not meant to be specialized for a specific domain type. Instead, specialize: interpolated = interpolate!(mgr, out, layout::Layout, refs, field, coord, increasing::Val)

Operators on Voronoi mesh

VoronoiSpheres.Stencils.Fix — Type

g = Fix(f, args)

Return callable g such that g(x,...) calls f by prepending args... before x...:

g(x...) == f(args..., x...)

This is similar to Base.Fix1, with several arguments.

VoronoiSpheres.Stencils.TRiSK — Method

vsphere = TRiSK(vsphere) # optional, returns only relevant fields as a named tuple
trisk = TRiSK(vsphere, edge, Val(N))
U_perp[edge]    = trisk(U)        # linear, single-layer, U::AbstractVector
U_perp[k, edge] = trisk(U, k)     # linear, multi-layer, U::AbstractMatrix
qU[edge]        = trisk(U, q)     # nonlinear, single-layer
qU[k, edge]     = trisk(U, q, k)  # nonlinear, multi-layer

Compute TRiSK operator U⟂ or q×U at edge::Int of vsphere::VoronoiSphere.

U is a contravariant vector field known at edges. U_perp is a covariant vector field known at edges.

N=sphere.trisk_deg[edge] is the number of edges involved in the TRiSK stencil and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into TRiSK and trisk.

VoronoiSpheres.Stencils.average_ev_form — Method

vsphere = average_ev_form(vsphere) # optional, returns only relevant fields as a named tuple
avg = average_ev(vsphere, dual)
qv[dual] = avg(qe)         # single-layer, qv::AbstractVector
qv[k, dual] = avg(qe, k)   # multi-layer, qv::AbstractMatrix

Estimate scalar field integrated over dual_cell::Int of vsphere::VoronoiSphere as the half-sum of values at edges.

qe is a density (two-form) over edges. To obtain a scalar, divide qe by the edge cell area vsphere.Ae qv is a density (two-form) over dual cells. To obtain a scalar, divide qv by the dual cell area vsphere.Av

@inbounds propagates into average_ev_form and avg.

VoronoiSpheres.Stencils.average_ie — Method

vsphere = average_ie(vsphere) # optional, returns only relevant fields as a named tuple
avg = average_ie(vsphere, edge)
qe[edge] = avg(qi)         # single-layer, qi::AbstractVector
qe[k, edge] = avg(qi, k)   # multi-layer, qi::AbstractMatrix

Interpolate scalar field at edge::Int of vsphere::VoronoiSphere by a centered average (second-order accurate).

qi is a scalar field known at primal cells. qe is a scalar field known at edges.

@inbounds propagates into average_ie and avg.

VoronoiSpheres.Stencils.average_iv — Method

vsphere = average_iv(vsphere) # optional, returns only relevant fields as a named tuple
avg = average_iv(vsphere, dual_cell)
qv[dual_cell] = avg(qi)         # single-layer, qi::AbstractVector
qv[k, dual_cell] = avg(qi, k)   # multi-layer, qi::AbstractMatrix

Estimate scalar field integrated over dual_cell::Int of vsphere::VoronoiSphere as an area-weighted sum of values sampled at primal cell centers.

qi is a scalar field known at primal cells. qv is a scalar field known at dual cells.

@inbounds propagates into average_iv and avg.

VoronoiSpheres.Stencils.average_iv_form — Method

vsphere = average_iv_form(vsphere) # optional, returns only relevant fields as a named tuple
avg = average_iv_form(vsphere, dual_cell)
qv[dual_cell] = avg(qi)         # single-layer, qi::AbstractVector
qv[k, dual_cell] = avg(qi, k)   # multi-layer, qi::AbstractMatrix

Interpolate scalar field at dual_cell::Int of vsphere::VoronoiSphere by an area-weighted sum (first-order accurate).

qi is a scalar field known at primal cells. qv is a density (two-form) over dual cells. To obtain a scalar, divide qv by the dual cell area vsphere.Av

@inbounds propagates into average_iv_form and avg.

VoronoiSpheres.Stencils.average_ve — Method

vsphere = average_ve(vsphere) # optional, returns only relevant fields as a named tuple
avg = average_ve(vsphere, edge)
qe[edge] = avg(qv)         # single-layer, qv::AbstractVector
qe[k, edge] = avg(qv, k)   # multi-layer, qv::AbstractMatrix

Interpolate scalar field at edge::Int of vsphere::VoronoiSphere by a centered average (first-order accurate).

qv is a scalar field known at dual cells. qe is a scalar field known at edges.

@inbounds propagates into average_ve and avg.

VoronoiSpheres.Stencils.average_vi_form — Method

vsphere = average_vi_form(vsphere) # optional, returns only relevant fields as a named tuple
avg = avg_vi_form(vsphere, cell, Val(N))
qi[cell] = avg(qv) # single-layer, qv::AbstractVector
qi[k, cell] = avg(qv, k)  # multi-layer, qv::AbstractMatrix

Estimate scalar field integrated over cell::Int of vsphere::VoronoiSphere as an area-weighted sum of values sampled at vertices. qv is a scalar field known at dual cells. qi is a density (two-form) over primal cells. To obtain a scalar, divide qi by the primal cell area vsphere.Ai

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into average_vi_form and avg.

VoronoiSpheres.Stencils.centered_flux — Method

vsphere = centered_flux(vsphere) # optional, returns only relevant fields as a named tuple
cflux = centered_flux(vsphere, edge)
flux[edge] = cflux(mass, ucov)         # single-layer, ucov::AbstractVector
flux[k, edge] = cflux(mass, ucov, k)   # multi-layer, ucov::AbstractMatrix

Compute centered flux at edge::Int of vsphere::VoronoiSphere, with respect to the unit sphere.

mass is a scalar field known at primal cells. ucov is a covariant vector field known at edges. flux is a contravariant vector field known at edges.

If ucov is defined with respect to a physical metric (e.g. in m²⋅s⁻¹) which is conformal, multiply cflux by the contravariant physical metric factor (in m⁻²). mass being e.g. in kg, on gets a flux in kg⋅s⁻¹ which can be fed into divergence.

@inbounds propagates into centered_flux and cflux.

VoronoiSpheres.Stencils.contraction — Method

vsphere = contraction(vsphere::VoronoiSphere) # optional, returns only relevant fields as a named tuple
contract = contraction(vsphere, cell::Int, v::Val{N})

# single-layer, `ucontra` and `vcov` are ::AbstractVector
uv[cell] = contract(ucontra, vcov) 

# multi-layer, `ucontra` and `vcov` are ::AbstractMatrix
uv[k, cell] = contract(ucontra, vcov, k)

Compute the contraction of ucov and vcov at cell::Int of vsphere::VoronoiSphere. ucontra is a contravariant vector field known at edges. vcov is a covariant vector field known at edges.

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into contraction and contract.

VoronoiSpheres.Stencils.cross_product — Method

vsphere = cross_product(vsphere) # optional, returns only relevant fields as a named tuple
cprod = cross_product(vsphere, edge, Val(N))
q[edge]    = cprod(U,V)        # single-layer, `U` and `V` are ::AbstractVector
q[k, edge] = cprod(U, V, k)    # multi-layer, `U` and `V` are ::AbstractMatrix

Compute the cross product U×V at edge::Int of vsphere::VoronoiSphere.

U and V are contravariant vector fields known at edges. q is a density (two-form) over edges. To obtain a scalar, divide q by the edge cell area vsphere.Ae

N=sphere.trisk_deg[edge] is the number of edges involved in the TRiSK stencil and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into cross_product and cprod.

VoronoiSpheres.Stencils.curl — Method

vsphere = curl(vsphere) # optional, returns only relevant fields as a named tuple
op = curl(vsphere, dual_cell)
curlu[dual_cell] = op(ucov)         # single-layer, ucov::AbstractVector
curlu[k, dual_cell] = op(ucov, k)   # multi-layer, ucov::AbstractMatrix

Compute curl of ucov at dual_cell::Int of vsphere::VoronoiSphere. ucov is a covariant vector field known at edges. curlu is a density (two-form) over dual cells. To obtain a scalar, divide curlu by the dual cell area vsphere.Av

@inbounds propagates into curl and op.

VoronoiSpheres.Stencils.div_centered_flux — Method

vsphere = div_centered_flux(vsphere) # optional, returns only relevant fields as a named tuple
div_flux = div_centered_flux(vsphere, cell)
divqF[cell] = div_flux(q, flux)         # single-layer, `flux` and `q` are ::AbstractVector
divqF[k, cell] = div_flux(q, flux, k)   # multi-layer, `flux` and `q` are ::AbstractMatrix

Compute divergence of q*flux at cell::Int of vsphere::VoronoiSphere, with respect to the unit sphere. q is interpolated by a simple centered average.

q is a scalar field known at primal cells. flux is a contravariant vector field known at edges. divqF is a density (two-form) over primal cells. To obtain a scalar, divide divqF by the primal cell area vsphere.Ai

@inbounds propagates into div_centered_flux and div_flux.

VoronoiSpheres.Stencils.div_form — Method

vsphere = div_form(vsphere) # optional, returns only relevant fields as a named tuple
divf = div_form(vsphere, cell, Val(N))
dvg[cell] = divf(flux) # single-layer, flux::AbstractVector
dvg[k, cell] = divf(flux, k)  # multi-layer, flux::AbstractMatrix

Compute divergence with respect to the unit sphere of flux at cell::Int of vsphere::VoronoiSphere. flux is a contravariant vector field known at edges. dvg is a density (two-form) over primal cells. To obtain a scalar, divide dvg by the primal cell area vsphere.Ai

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into div_form and divf.

VoronoiSpheres.Stencils.divergence — Method

vsphere = divergence(vsphere) # optional, returns only relevant fields as a named tuple
div = divergence(vsphere, cell, Val(N))
dvg[cell] = div(flux) # single-layer, flux::AbstractVector
dvg[k, cell] = div(flux, k)  # multi-layer, flux::AbstractMatrix

Compute divergence with respect to the unit sphere of flux at cell::Int of vsphere::VoronoiSphere. flux is a contravariant vector field known at edges. dvg is a scalar field known at primal cells.

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into divergence and div.

VoronoiSpheres.Stencils.dot_grad — Method

vsphere = dot_grad(vsphere) # optional, returns only relevant fields as a named tuple
dotgrad = dot_grad(vsphere, cell)
Fgradq[cell] = dotgrad(flux, q)         # `U` is a *contravariant* vector field known at edges.
Fgradq[k, cell] = dotgrad(flux, q, k)   # multi-layer, flux::AbstractMatrix

At cell::Int, compute the dot product of flux with the gradient of q with respect to the unit sphere .

q is a scalar field known at primal cells. flux is a contravariant vector field known at edges. Fgradq is a density (two-form) over primal cells. To obtain a scalar, divide Fgradq by the primal cell area vsphere.Ai

@inbounds propagates into dot_grad and dotgrad.

VoronoiSpheres.Stencils.dot_prod_contra — Method

vsphere = dot_prod_contra(vsphere::VoronoiSphere) # optional, returns only relevant fields as a named tuple
dot_prod = dot_prod_contra(vsphere, cell::Int, v::Val{N})

# single-layer, `U` and `V` are ::AbstractVector
dp[cell] = dot_prod(U, V) 

# multi-layer, `U` and `V` are ::AbstractMatrix
dp[k, cell] = dot_prod(U, V, k)

Compute dot product with respect to the unit sphere of U, V at cell::Int of vsphere::VoronoiSphere. U and V are contravariant vector fields known at edges.

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into dot_prod_contra and dot_prod.

VoronoiSpheres.Stencils.dot_product — Method

vsphere = dot_product(vsphere::VoronoiSphere) # optional, returns only relevant fields as a named tuple
dot_prod = dot_product(vsphere, cell::Int, v::Val{N})

# single-layer, `ucov` and `vcov` are ::AbstractVector
dp[cell] = dot_prod(ucov, vcov) 

# multi-layer, `ucov` and `vcov` are ::AbstractMatrix
dp[k, cell] = dot_prod(ucov, vcov, k)

Compute dot product with respect to the unit sphere of ucov, vcov at cell::Int of vsphere::VoronoiSphere. ucov and vcov are covariant vector fields known at edges.

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into dot_product and dot_prod.

VoronoiSpheres.Stencils.dot_product_form — Method

vsphere = dot_product_form(vsphere::VoronoiSphere) # optional, returns only relevant fields as a named tuple
dot_prod = dot_product_form(vsphere, cell::Int, v::Val{N})

# single-layer, `ucov` and `vcov` are ::AbstractVector
dp[cell] = dot_prod(ucov, vcov) 

# multi-layer, `ucov` and `vcov` are ::AbstractMatrix
dp[k, cell] = dot_prod(ucov, vcov, k)

Compute dot product with respect to the unit sphere of ucov, vcov at cell::Int of vsphere::VoronoiSphere. ucov and vcov are covariant vector fields known at edges. dp is a density (two-form) over primal cells. To obtain a scalar, divide dp by the primal cell area vsphere.Ai

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into dot_product_form and dot_prod.

VoronoiSpheres.Stencils.grad_form — Method

vsphere = grad_form(vsphere) # optional, returns only relevant fields as a named tuple
grad = grad_form(vsphere, edge)
gradcov[edge] = grad(Q)         # single-layer, Q::AbstractVector
gradcov[k, edge] = grad(Q, k)   # multi-layer, Q::AbstractMatrix

Compute gradient of Q at edge::Int of vsphere::VoronoiSphere.

Q is a density (two-form) over primal cells. To obtain a scalar, divide Q by the primal cell area vsphere.Ai gradcov is a covariant vector field known at edges. gradcov is numerically zero-curl.

@inbounds propagates into grad_form and gradcov.

VoronoiSpheres.Stencils.gradient — Method

vsphere = gradient(vsphere) # optional, returns only relevant fields as a named tuple
grad = gradient(vsphere, edge)
gradcov[edge] = grad(q)         # single-layer, q::AbstractVector
gradcov[k, edge] = grad(q, k)   # multi-layer, q::AbstractMatrix

Compute gradient of q at edge::Int of vsphere::VoronoiSphere.

q is a scalar field known at primal cells. gradcov is a covariant vector field known at edges. gradcov is numerically zero-curl.

@inbounds propagates into gradient and gradcov.

VoronoiSpheres.Stencils.gradient3d — Method

vsphere = gradient3d(vsphere) # optional, returns only relevant fields as a named tuple
grad = gradient3d(vsphere, cell, Val(N))
gradq[ij] = grad(q)         # single-layer, q::AbstractVector
gradq[k, ij] = grad(q, k)   # multi-layer, q::AbstractMatrix

Compute 3D gradient of q at cell::Int of vsphere::VoronoiSphere. q is a scalar field known at primal cells. gradq is a 3D vector field yielding a 3-uple at each primal cell.

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into gradient3d and grad.

VoronoiSpheres.Stencils.gradperp — Method

vsphere = gradperp(vsphere) # optional, returns only relevant fields as a named tuple
grad = gradperp(vsphere, edge)
flux[edge] = grad(psi)         # single-layer, q::AbstractVector
flux[k, edge] = grad(psi, k)   # multi-layer, q::AbstractMatrix

Compute grad⟂ of streamfunction psi at edge::Int of vsphere::VoronoiSphere.

psi is a scalar field known at dual cells. flux is a contravariant vector field known at edges. flux is numerically non-divergent.

@inbounds propagates into gradperp and grad.

VoronoiSpheres.Stencils.mul_grad — Method

vsphere = mul_grad(vsphere) # optional, returns only relevant fields as a named tuple
mulgrad = mul_grad(vsphere, edge)
a∇b[edge] = mulgrad(a,b)     # single-layer, `a` and `b` are ::AbstractVector
a∇b[k, edge] = mulgrad(a, b, k)   # multi-layer, `a` and `b` are ::AbstractMatrix

Compute gradient of b at edge::Int of vsphere::VoronoiSphere, multiplied by a.

a and a are a scalar fields known at primal cells. a∇b is a covariant vector field known at edges.

@inbounds propagates into mul_grad and mulgrad.

VoronoiSpheres.Stencils.perp — Method

op = perp(vsphere) # optional, returns only relevant fields as a named tuple
op = perp(vsphere, ij)
U_perp[ij] = op(U)              # single-layer, U::AbstractVector
U_perp[k, ij] = op(U, k::Int)   # multi-layer, U::AbstractMatrix

Compute the perp operator U⟂ at edge::Int of vsphere::VoronoiSphere. Unlike the TRiSK operator, this operator is not antisymmetric but it has a smaller stencil and is numerically consistent.

Array U represents a vector field U by its components normal to edges of primal cells. U_perp represents similarly U⟂. Equivalently, it represents U by its components normal to edges of dual cells.

@inbounds propagates into perp and op.

VoronoiSpheres.Stencils.squared_covector — Method

vsphere = squared_covector(vsphere::VoronoiSphere) # optional, returns only relevant fields as a named tuple
square = squared_covector(vsphere, cell::Int, v::Val{N})

# single-layer, ucov::AbstractVector
u_squared_form[cell] = square(ucov) 

# multi-layer, ucov::AbstractMatrix
u_squared[k, cell] = square(ucov, k)

Compute dot product with respect to the unit sphere of ucov and istelf at cell::Int of vsphere::VoronoiSphere. ucov is a covariant vector field known at edges. u_squared is a density (two-form) over primal cells. To obtain a scalar, divide u_squared by the primal cell area vsphere.Ai

N=sphere.primal_deg[cell] is the number of cell edges and must be provided as a compile-time constant for performance. This may be done via the macro @unroll from ManagedLoops.

@inbounds propagates into squared_covector and square.

VoronoiSpheres.VoronoiOperators.AsDensity — Method

as_density = AsDensity(vsphere) # a `LazyDiagonalOp`
density = as_density(scalar)    # a `WritableDVP` (diagonal-vector-product)
op!(density, ...)               # pass `density as *output* argument

Given a zero-form scalar, as_density returns the equivalent two-form as a lazy, write-only AbstractArray to be passed to a VoronoiOperator op! as an output argument.

VoronoiSpheres.VoronoiOperators.pdv — Function

fa = pdv(fun, a)
fa, fb = pdv(fun, a, b)
fa, fb, fc = pdv(fun, a, b, c)

Return the partial derivatives of scalar function fun evaluated at input a, .... This function is implemented only when the package ForwardDiff is loaded either directly from the main program or via some dependency.

Virtual zero-filled arrays

VoronoiSpheres.ZeroArrays.zero_array — Method

z = zero_array(a::AbstractArray)

Return z which behaves like a read-only array with the same axes as a, filled with zeros.

z[i...] does not check bounds and always returns the special value Zero(), which behaves like 0 when passed to +, * and muladd.